speech_to_text_streamlit_project/requirements.txt
tensorflow-cpu==2.13.0
numpy<1.24
librosa
soundfile
sounddevice
pandas
streamlit